@ feat(exercicios): unifica Cardio+Peso Corporal no tipo Performance com métrica/unidade livre
Substitui os tipos CARDIO e PESO_CORPORAL por um único PERFORMANCE, em que o personal
define uma métrica numérica livre, sua unidade (≤7 chars, combobox com presets) e a
direção da evolução (MAIOR = mais é melhor; MENOR = menos é melhor, p/ tempo/pace).
A métrica alimenta o gráfico de evolução e o PR (melhor série conforme a direção).

Backend:
- enums: novo PERFORMANCE + normalizar_tipo_exercicio() (compat de leitura p/ legados)
- novo campo metrica_direcao em Exercicio/Template/Pacote/TreinoExport/Sessao
- SerieExec.reps relaxado para float (suporta km/min/decimais)
- _calcular_pr e evolucao_exercicio: ramo PERFORMANCE (max/min via update_if_greater/less)
- unidade do PR propagada ao card de checkin
- pacote_service/treinos: propagam unidade_reps e metrica_direcao no import/export
- script migrar_resistencia.py: CARDIO/PESO_CORPORAL -> PERFORMANCE (idempotente)

Frontend:
- TipoExercicio = FORCA|PERFORMANCE + normalizeTipoExercicio() p/ dados legados
- form (aluno e template): seletor de unidade (datalist) + direção; input único no app do aluno
- evolução/PR/histórico/checkin: rótulos por unidade e direção
- prompts de IA e ajuda atualizados p/ FORCA|PERFORMANCE

@

+++ backend/app/models/registro.py
@@ -8,7 +8,7 @@ from app.models.enums import Ator, CanalOrigem, Classificacao
 class SerieExec(BaseModel):
     """Uma série executada. Acumuladas numa lista do mesmo Registro (ESPEC §3)."""
     carga: Optional[str] = None
-    reps: Optional[int] = None
+    reps: Optional[float] = None               # repetições (FORCA) ou métrica livre (PERFORMANCE — aceita decimais: km, min…)
     rpe: Optional[float] = None                # percepção de esforço (0-10)
